+++ packages/plugin-store/src/index.ts
@@ -119,7 +119,7 @@ export default async (api) => {
 
   gen.render();
   onHook('before.start.run', async () => {
-    applyMethod('watchFileChange', /models\/.*|model.*|pages\/\w+\/index(.jsx?|.tsx)/, () => {
+    applyMethod('watchFileChange', /models\/.*|model.*|store.*|pages\/\w+\/index(.jsx?|.tsx)/, () => {
       gen.render();
     });
   });